Изменена структура проекта
