+++ misc/states.py
@@ -1,5 +1,6 @@
 from aiogram.dispatcher.filters.state import StatesGroup, State
 
+
 class CreateDocument(StatesGroup):
     Processor = State()
     UseYandexGPT = State()
