+++ requirements.txt
@@ -4,3 +4,10 @@ python-docx==1.1.0
 openpyxl~=3.1.2
 reportlab~=4.2.0
 docx2pdf~=0.1.8
+
+environs==11.0.0
+aiogram==2.25.2
+gino==1.0.1
+SQLAlchemy==1.3.24
+async-timeout==4.0.3
+asyncpg==0.29.0
