+++ utils/db_commands.py
@@ -1,4 +1,4 @@
-from bot.utils.database import User
+from utils.models import User
 from aiogram import types
 
 async def get_user():
